requirements.txt
futures==3.0.3
goslate==1.4.0
phpsh==1.3
py==1.4.30
pycurl==7.19.5.1
pyenchant==1.6.6
pyTelegramBotAPI==0.2.9
pytest==2.7.2
python-telegram-bot==2.4
requests==2.7.0
six==1.9.0
wheel==0.24.0
